service creation in db

+++ src/database/priorityQueue.ts
@@ -1,24 +1,24 @@
-import { execFileSync } from "node:child_process";
 import { db } from "./index"
 
+// Inserção
 export function insertInPriorityQueue(
-  id: number,
-  patientId: number,
-  riskRatingId: number
+  atendimento_id: number,
+  paciente_id: number,
+  classificacao_risco_id: number
 ) {
   const stmt = db.prepare(`
     INSERT INTO FilaDePrioridade (
       atendimento_id, paciente_id, 
       classificacao_risco_id
     ) VALUES (?, ?, ?)
-  `)
+  `);
 
   const info = stmt.run(
-    id,
-    patientId,
-    riskRatingId
-  )
-  return info.lastInsertRowid
+    atendimento_id,
+    paciente_id,
+    classificacao_risco_id
+  );
+  return info.lastInsertRowid;
 }
 
 export function getPriorityQueue() {
@@ -28,28 +28,27 @@ export function getPriorityQueue() {
   return stmt.all();
 }
 
-export function getPriorityQueueByCpf(cpf: string) {
+export function deletePatientFromPriorityQueue(paciente_id: number) {
   const stmt = db.prepare(`
-    'SELECT * FROM FilaDePrioridade 
-    WHERE cpf = @cpf'
+    DELETE FROM FilaDePrioridade WHERE paciente_id = ?
   `);
-  return stmt.get({ cpf });
-  
-}
-
-export function deletePatientFromPriorityQueue(patientId: number) {
-  const stmt = db.prepare('DELETE FROM FilaDePrioridade WHERE paciente_id = ?')
-  const info = stmt.run(patientId)
-  return info.changes
+  const info = stmt.run(paciente_id);
+  return info.changes;
 }
 
 export function deletePatientFromPriorityQueueByCpf(cpf: string) {
+  const paciente = db.prepare(`
+    SELECT id FROM Paciente WHERE cpf = ?
+  `).get(cpf) as { id: number } | undefined;
+
+  if (!paciente) {
+    throw new Error("Paciente com o CPF informado não encontrado.");
+  }
+
   const stmt = db.prepare(`
-    DELETE FROM FilaDePrioridade 
-    WHERE paciente_id = (
-      SELECT id FROM Paciente WHERE cpf = ?
-    )
-  `)
-  const info = stmt.run(cpf)
-  return info.changes
+    DELETE FROM FilaDePrioridade WHERE paciente_id = ?
+  `);
+  const info = stmt.run(paciente.id);
+  return info.changes;
 }
+
